requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pytrends>=4.9.0
praw>=7.7.0
newsapi-python>=0.2.7
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.2.0
xgboost>=1.7.0
prophet>=1.1.4
fastapi>=0.100.0
uvicorn>=0.22.0
streamlit>=1.25.0
boto3>=1.28.0
requests>=2.31.0
joblib>=1.3.0
spacy>=3.6.0
